requirements.txt
Django==1.7.1
coverage==3.7.1
dj-database-url==0.3.0
dj-static==0.0.6
django-toolbelt==0.0.1
factory-boy==2.4.1
gunicorn==18.0
mock==1.0.1
psycopg2==2.5.1
py-bcrypt==0.4
pystache==0.5.4
static==0.4
wsgiref==0.1.2
unittest2==0.8.0
